chore: move carriers methods in carrierRepository
