+++ src/Adapter/Product/Repository/ProductRepository.php
@@ -970,85 +970,6 @@ class ProductRepository extends AbstractMultiShopObjectModelRepository
     }
 
     /**
-     * @return array<int, array{id_carrier: int|string, name: string}>
-     */
-    public function getCarriersByProductId(ProductId $productId, ShopId $shopId): array
-    {
-        $hasRestrictions = $this->hasCarrierRestrictions($productId->getValue(), $shopId->getValue());
-
-        if (!$hasRestrictions) {
-            return $this->getAllActiveCarriers($shopId->getValue());
-        }
-
-        return $this->getRestrictedCarriers($productId->getValue(), $shopId->getValue());
-    }
-
-    private function hasCarrierRestrictions(int $productId, int $shopId): bool
-    {
-        $qb = $this->connection->createQueryBuilder();
-
-        $count = $qb->select('COUNT(*)')
-            ->from($this->dbPrefix . 'product_carrier', 'pc')
-            ->where('pc.id_product = :productId')
-            ->andWhere('pc.id_shop = :shopId')
-            ->setParameter('productId', $productId)
-            ->setParameter('shopId', $shopId)
-            ->executeQuery()
-            ->fetchOne();
-
-        return $count > 0;
-    }
-
-    /**
-     * @return array<int, array{id_carrier: int|string, name: string}>
-     */
-    private function getAllActiveCarriers(int $shopId): array
-    {
-        $qb = $this->connection->createQueryBuilder();
-
-        return $qb->select(
-            'c.id_carrier',
-            'c.name'
-        )
-            ->from($this->dbPrefix . 'carrier', 'c')
-            ->where('c.deleted = 0')
-            ->andWhere('c.active = 1')
-            ->orderBy('c.name', 'ASC')
-            ->setParameter('shopId', $shopId)
-            ->executeQuery()
-            ->fetchAllAssociative();
-    }
-
-    /**
-     * @return array<int, array{id_carrier: int|string, name: string}>
-     */
-    private function getRestrictedCarriers(int $productId, int $shopId): array
-    {
-        $qb = $this->connection->createQueryBuilder();
-
-        return $qb->select(
-            'c.id_carrier',
-            'c.name'
-        )
-            ->from($this->dbPrefix . 'product_carrier', 'pc')
-            ->innerJoin(
-                'pc',
-                $this->dbPrefix . 'carrier',
-                'c',
-                'pc.id_carrier_reference = c.id_reference'
-            )
-            ->where('pc.id_product = :productId')
-            ->andWhere('pc.id_shop = :shopId')
-            ->andWhere('c.deleted = 0')
-            ->andWhere('c.active = 1')
-            ->orderBy('c.name', 'ASC')
-            ->setParameter('productId', $productId)
-            ->setParameter('shopId', $shopId)
-            ->executeQuery()
-            ->fetchAllAssociative();
-    }
-
-    /**
      * @todo: this should be removable soon once the deprecated stock properties have been removed see PR #26682
      *
      * @param Product $product
